POST and GET typecast updates. Fred

+++ modules/administrationmodule/actions/session_kickuser.php
@@ -21,9 +21,9 @@
 
 if (!defined('PATHOS')) exit('');
 
-if (pathos_permissions_check('user_management',pathos_core_makeLocation('administrationmodule'))) {
+if (isset($_GET['id']) && pathos_permissions_check('user_management',pathos_core_makeLocation('administrationmodule'))) {
 	if (!defined('SYS_USERS')) require_once(BASE.'subsystems/users.php');
-	$u = pathos_users_getUserById($_GET['id']);
+	$u = pathos_users_getUserById(intval($_GET['id']));
 	if ($u->is_acting_admin == 0 || ($user->is_admin == 1 && $u->is_admin == 0)) {
 		$db->delete('sessionticket','uid='.$u->id);
 	}
